pyproject.toml
[project]
name = "stencil-maker"
version = "0.1.0"
description = "Generate an image as a text stencil"
authors = [
    { name = "mamg22", email = "45301823+mamg22@users.noreply.github.com" }
]
dependencies = [
    "pillow>=10.3.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.optional-dependencies]
server = [
    "flask>=3.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/stencil_maker"]
